refactor : 카테고리 jpa 적용
