+++ src/main/java/techcourse/myblog/domain/User.java
@@ -7,11 +7,12 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
-
-    private String name;
-    //@Column(name = "email", unique = true)
+    @Column(unique = true)
     private String email;
+
     private String password;
+    private String name;
+
 
     public User() {}
 
@@ -25,32 +26,16 @@ public class User {
         return id;
     }
 
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
     public String getEmail() {
         return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
     }
 
     public String getPassword() {
         return password;
     }
 
-    public void setPassword(String password) {
-        this.password = password;
+    public String getName() {
+        return name;
     }
 
     @Override
