+++ src/main/java/techcourse/myblog/domain/UserDto.java
@@ -56,8 +56,6 @@ public class UserDto {
         this.password = password;
     }
 
-
-
     @Override
     public String toString() {
         return "UserDto{" +
