+++ src/main/java/techcourse/myblog/web/CategoryController.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import techcourse.myblog.domain.ArticleRepositoryImpl;
 import techcourse.myblog.domain.Category;
-import techcourse.myblog.domain.CategoryRepository;
+import techcourse.myblog.domain.CategoryRepositoryImpl;
 
 @Controller
 public class CategoryController {
@@ -16,11 +16,11 @@ public class CategoryController {
     private ArticleRepositoryImpl articleRepositoryImpl;
 
     @Autowired
-    private CategoryRepository categoryRepository;
+    private CategoryRepositoryImpl categoryRepositoryImpl;
 
     @PostMapping("/categories/add")
     public String addCategories(Category category) {
-        categoryRepository.addCategory(category);
+        categoryRepositoryImpl.addCategory(category);
 
         return "redirect:/";
     }
@@ -31,7 +31,7 @@ public class CategoryController {
         category.setCategoryId(categoryId);
 
         if (articleRepositoryImpl.findByCategoryId(categoryId).isEmpty()) {
-            categoryRepository.delete(category);
+            categoryRepositoryImpl.delete(category);
         }
 
         return "redirect:/";
