+++ src/main/resources/application.properties
@@ -11,11 +11,13 @@ spring.devtools.livereload.enabled=true
 #spring.datasource.password=password
 #spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 
-#spring.jpa.hibernate.ddl-auto=create-drop
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create-drop
+#spring.jpa.hibernate.ddl-auto=update
 spring.datasource.url=jdbc:mysql://localhost:3306/db_myblog?serverTimezone=UTC&useSSL=false
 spring.datasource.username=user
 spring.datasource.password=gmlgus12
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialect
+
 
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql= true
